backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
httpx==0.28.1
requests==2.32.3
sse-starlette==2.2.1
google-generativeai==0.8.3
shapely==2.0.6
geopy==2.4.1
pandas==2.2.3
numpy>=1.24,<2.1
python-multipart==0.0.20
